requirements.txt
allure-pytest==2.8.31
allure-python-commons==2.8.31
apipkg==1.5
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.4
common==0.1.2
configparser==5.0.2
cryptography==3.4.7
ddmctest==1.0.0
decorator==5.0.9
elasticsearch==7.12.0
et-xmlfile==1.1.0
execnet==1.8.1
Flask==1.0.2
greenlet==1.1.0

idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.10
json-tools==0.4.1
jsonpath==0.82
jsonschema==3.2.0
logger==1.4
loguru==0.5.3
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.20.3
openpyxl==3.0.7
packaging==20.8
paramiko==2.7.2
phoenixdb==1.0.1
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycparser==2.20
pydantic==1.8.1
pandas==1.3.1
pymongo==3.11.3
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pysolr==3.9.0
pytest==6.2.1
pytest-assume==2.4.2
pytest-dependency==0.5.1
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-repeat==0.9.1
pytest-report==0.2.1
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.1
pytest_check==1.0.1
PyTestReport==0.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytils==0.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-gssapi==1.2.3
requests-toolbelt==0.9.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.18
sshtunnel==0.4.0
toml==0.10.2
tools==0.1.9
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
win32-setctime==1.0.3
xlrd==2.0.1
zipp==3.4.0
wheel==0.36.2
setuptools==57.4.0
regex==2021.4.4
redis==2.10.6
redis-py-cluster==1.3.6
python-jenkins==1.7.0
keywords==0.0.1
multi-key-dict==2.0.3
nltk==3.6.1
pbr==5.6.0
pip==21.2.3
data==0.4
funcsigs==1.0.2
gssapi==1.6.14
joblib==1.0.1
charset-normalizer==2.0.4
chevron==0.14.0